Create a separate robozonky-test Maven module
