+++ robozonky-app/src/test/java/com/github/robozonky/app/AbstractEventsAndStateLeveragingTest.java
@@ -16,8 +16,8 @@
 
 package com.github.robozonky.app;
 
-import com.github.robozonky.common.AbstractStateLeveragingTest;
 import com.github.robozonky.internal.api.Settings;
+import com.github.robozonky.test.AbstractStateLeveragingTest;
 import org.junit.After;
 import org.junit.Rule;
 import org.junit.contrib.java.lang.system.ProvideSystemProperty;
