+++ robozonky-test/src/main/java/com/github/robozonky/test/IoTestUtil.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.github.robozonky.util;
+package com.github.robozonky.test;
 
 import java.io.File;
 import java.io.IOException;
